testando

+++ PI3/Entidades/Pets.cs
@@ -9,7 +9,7 @@ namespace PI3.Entidades
         public int Id { get; set; }
         public string NOME { get; set; }
         public string RACA { get; set; }
-        public decimal PESO { get; set; }
+        public int PESO { get; set; }
         public DateTime DATANASCIMENTO { get; set; }
         public string DESCRICAO { get; set; }
         public string CaminhoImagem { get; set; }
